Fix a bug resolving a URL with the relative URL "//", which would generate an invalid host name component (it wouldn't get set).
This also exposed some bugs in our host handling (hosts were special cased so
when they are empty, they had a length of 0, and we weren't always doing this
properly). I've been wanting to make host handling more consistent, so I did
it in this change. I also used the reset() function more instead of assigning
to the empty object (this function is newer than most of the code).