Fix several limitations of the canonicalizer. This validates characters in the reference fragment and user info sections. I did some cleanup associated with this for character validity checking. This also makes the query escape certain characters like # which could possibly get set if the query is replaced with something new.
This is a major change to the host canonicalizer. We now support a third mode
for compatibility to allow more characters and to escape them if we see them
in hosts. Even though these are not valid DNS names, I have seen them being
used in third-party protocols that happen to be "standard" and most programs
support them.