Update the unit tests to allow a colon in UNC hostnames on Windows. This test was broken in r107 but we weren't running the tests properly so never noticed.
Colon is probably not allowed by Windows, but its the "natural" result of the
way the canonicalizer is written. There doesn't seem to be a strong argument
for why allowing it here would be bad, so we just tolerate it and the load will
fail later.
Review URL: http://codereview.chromium.org/173535