Fix encoding of the query part of an URL.
Encoding of the query part of an URL was inconsistent:
If it was part of a larger relative URL, it would get encoded.
However, if the query part was stand-alone, then it would not be encoded.

As indicated by the original layout test referenced in the bug (see below),
this patch changes the code so that the query part is encoded in both cases.

Note, however, that with this patch the layout test will NOT yet fully pass,
since it also handles the fragment part differently.
As the different handling of the fragment part seems to be quite on purpose,
I have not modified the behavior - see the discussion at

http://code.google.com/p/chromium/issues/detail?id=20507 .


BUG=20507, 8912
TEST=LayoutTests/http/tests/uri/resolve-encoding-relative.html
Patch by Roland Steiner, review URL http://codereview.chromium.org/243028